feat: UI/UX改善とナビゲーション再構築
- サイドバーを半透明化し、後ろのコンテンツが見えるように改善
- メインコンテンツエリアの背景色を明るいグレーに変更
- オーバーレイを透明化して視認性を向上
- 設定ページから収入予測・シミュレーション機能を削除
- 初回ログイン画面とオンボーディングの完全実装
- APIエラーの修正（Supabase環境変数問題の一時的回避）
- ダッシュボードページの初期状態表示を改善
- 勤務管理機能の統合とページ構造の最適化
