+++ src/app/(dashboard)/layout.tsx
@@ -1,6 +1,7 @@
 "use client";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
+import { useState, useEffect } from "react";
 import { 
   HomeIcon, 
   ClockIcon, 
@@ -9,16 +10,19 @@ import {
   PlusIcon, 
   ChartBarIcon,
   UserIcon,
-  Cog6ToothIcon
+  Cog6ToothIcon,
+  Bars3Icon,
+  XMarkIcon,
+  CurrencyYenIcon,
+  CalendarIcon
 } from "@heroicons/react/24/outline";
 
+// 学生目線でのシンプルなナビゲーション
 const navigation = [
-  { name: 'ダッシュボード', href: '/', icon: HomeIcon },
-  { name: '勤務管理', href: '/schedule', icon: ClockIcon },
-  { name: 'シミュレーション', href: '/simulation', icon: CalculatorIcon },
-  { name: '収入予測', href: '/prediction', icon: ArrowTrendingUpIcon },
-  { name: '収入記録', href: '/income', icon: PlusIcon },
-  { name: '詳細レポート', href: '/profile', icon: ChartBarIcon },
+  { name: 'ホーム', href: '/dashboard', icon: HomeIcon, description: '現在の状況とクイックアクション' },
+  { name: '収入記録', href: '/income', icon: CurrencyYenIcon, description: '実際の収入を記録・管理' },
+  { name: '勤務管理', href: '/work', icon: ClockIcon, description: '定期勤務とシフト管理' },
+  { name: '設定・予測', href: '/profile', icon: UserIcon, description: '個人設定・収入予測・シミュレーション' },
 ];
 
 export default function DashboardLayout({
@@ -27,20 +31,46 @@ export default function DashboardLayout({
   children: React.ReactNode;
 }) {
   const pathname = usePathname();
+  const [sidebarOpen, setSidebarOpen] = useState(false);
+  const [isLoading, setIsLoading] = useState(true);
+
+  // 初期化時のローディング状態
+  useEffect(() => {
+    const timer = setTimeout(() => {
+      setIsLoading(false);
+    }, 500);
+    return () => clearTimeout(timer);
+  }, []);
 
   return (
     <div className="min-h-screen bg-gray-50">
+      {/* オーバーレイ - 完全に透明 */}
+      {sidebarOpen && (
+        <div 
+          className="fixed inset-0 z-40"
+          onClick={() => setSidebarOpen(false)}
+        />
+      )}
+
       {/* サイドバー */}
-      <div className="fixed inset-y-0 left-0 z-50 w-64 bg-white shadow-lg">
+      <div className={`fixed top-0 left-0 z-50 w-64 h-full bg-white/80 backdrop-blur-sm shadow-lg transform transition-transform duration-300 ease-in-out ${
+        sidebarOpen ? 'translate-x-0' : '-translate-x-full'
+      }`}>
         <div className="flex h-full flex-col">
-          {/* ロゴ */}
-          <div className="flex h-16 items-center justify-center border-b border-gray-200">
+          {/* ヘッダー */}
+          <div className="flex h-16 items-center justify-between px-4 border-b border-gray-200">
             <Link href="/" className="flex items-center gap-2">
               <div className="h-8 w-8 rounded-lg bg-blue-600 flex items-center justify-center">
                 <span className="text-white font-bold text-sm">MS</span>
               </div>
               <span className="font-semibold text-gray-900">Money Student</span>
             </Link>
+            <button
+              onClick={() => setSidebarOpen(false)}
+              className="p-1 rounded-md text-gray-400 hover:text-gray-600"
+            >
+              <XMarkIcon className="h-6 w-6" />
+            </button>
           </div>
 
           {/* ナビゲーション */}
@@ -51,14 +81,18 @@ export default function DashboardLayout({
                 <Link
                   key={item.name}
                   href={item.href}
-                  className={`group flex items-center gap-3 rounded-lg px-3 py-2 text-sm font-medium transition-colors ${
+                  onClick={() => setSidebarOpen(false)}
+                  className={`group flex items-center gap-3 rounded-lg px-3 py-3 text-sm font-medium transition-colors ${
                     isActive
                       ? 'bg-blue-50 text-blue-700 border-r-2 border-blue-700'
                       : 'text-gray-700 hover:bg-gray-100 hover:text-gray-900'
                   }`}
                 >
-                  <item.icon className="h-5 w-5" />
-                  {item.name}
+                  <item.icon className="h-5 w-5 flex-shrink-0" />
+                  <div className="flex-1">
+                    <div>{item.name}</div>
+                    <div className="text-xs text-gray-500 font-normal">{item.description}</div>
+                  </div>
                 </Link>
               );
             })}
@@ -86,9 +120,30 @@ export default function DashboardLayout({
       </div>
 
       {/* メインコンテンツ */}
-      <div className="pl-64">
-        <main className="min-h-screen">
-          {children}
+      <div>
+        {/* ヘッダー */}
+        <div className="bg-white border-b border-gray-200 px-4 py-3">
+          <div className="flex items-center justify-between">
+            <button
+              onClick={() => setSidebarOpen(true)}
+              className="p-2 rounded-md text-gray-400 hover:text-gray-600"
+            >
+              <Bars3Icon className="h-6 w-6" />
+            </button>
+            <Link href="/" className="flex items-center gap-2">
+              <div className="h-6 w-6 rounded-lg bg-blue-600 flex items-center justify-center">
+                <span className="text-white font-bold text-xs">MS</span>
+              </div>
+              <span className="font-semibold text-gray-900 text-sm">Money Student</span>
+            </Link>
+            <div className="w-10"></div> {/* 中央揃えのためのスペーサー */}
+          </div>
+        </div>
+
+        <main className="min-h-screen bg-gray-100 p-4 lg:p-6">
+          <div className="max-w-7xl mx-auto">
+            {children}
+          </div>
         </main>
       </div>
     </div>
