+++ src/app/api/profile/route.ts
@@ -1,42 +1,12 @@
 import { NextResponse } from 'next/server';
-import { createSupabaseServerClient } from '@/lib/supabaseServer';
 
+// 一時的にSupabaseを無効にして、ローカルストレージのみで動作
 export async function GET() {
-  const supabase = createSupabaseServerClient();
-  // 認証ユーザーのプロフィールを1件取得
-  const { data: { user }, error: userError } = await supabase.auth.getUser();
-  if (userError || !user) {
-    return NextResponse.json({ error: 'unauthorized' }, { status: 401 });
-  }
-
-  const { data, error } = await supabase
-    .from('profiles')
-    .select('*')
-    .eq('id', user.id)
-    .single();
-
-  if (error) return NextResponse.json({ error: error.message }, { status: 500 });
-  return NextResponse.json({ profile: data });
+  // ローカルストレージのデータを返す（クライアントサイドで処理）
+  return NextResponse.json({ profile: null });
 }
 
 export async function PUT(req: Request) {
-  const body = await req.json().catch(() => ({}));
-  const supabase = createSupabaseServerClient();
-  const { data: { user }, error: userError } = await supabase.auth.getUser();
-  if (userError || !user) {
-    return NextResponse.json({ error: 'unauthorized' }, { status: 401 });
-  }
-
-  const patch = body?.profile ?? {};
-  const upsert = { id: user.id, ...patch };
-
-  const { data, error } = await supabase
-    .from('profiles')
-    .upsert(upsert)
-    .eq('id', user.id)
-    .select('*')
-    .single();
-
-  if (error) return NextResponse.json({ error: error.message }, { status: 500 });
-  return NextResponse.json({ profile: data });
+  // ローカルストレージのデータを更新（クライアントサイドで処理）
+  return NextResponse.json({ success: true });
 }
